JAMES-2586 Remove james-server-cassandra-app direct dependency (#1875)
