+++ mpt/impl/smtp/cassandra-rabbitmq-object-storage/pom.xml
@@ -99,6 +99,13 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-guice-opensearch</artifactId>
+            <version>${project.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-util</artifactId>
             <type>test-jar</type>
             <scope>test</scope>
