+++ server/container/guice/opensearch/src/test/java/org/apache/james/DockerOpenSearchRule.java
@@ -17,45 +17,39 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.modules;
+package org.apache.james;
 
-import org.apache.james.CleanupTasksPerformer;
 import org.apache.james.backends.opensearch.DockerOpenSearch;
-import org.apache.james.backends.opensearch.OpenSearchConfiguration;
+import org.apache.james.backends.opensearch.DockerOpenSearchSingleton;
+import org.apache.james.modules.mailbox.TestDockerOpenSearchModule;
+import org.junit.runner.Description;
+import org.junit.runners.model.Statement;
 
-import com.google.inject.AbstractModule;
-import com.google.inject.multibindings.Multibinder;
+import com.google.inject.Module;
 
-public class TestDockerOpenSearchModule extends AbstractModule {
+public class DockerOpenSearchRule implements GuiceModuleTestRule {
+    private final DockerOpenSearch openSearch = DockerOpenSearchSingleton.INSTANCE;
 
-    private static class ESContainerCleanUp implements CleanupTasksPerformer.CleanupTask {
-
-        private final DockerOpenSearch elasticSearch;
-
-        private ESContainerCleanUp(DockerOpenSearch elasticSearch) {
-            this.elasticSearch = elasticSearch;
-        }
-
-        @Override
-        public Result run() {
-            elasticSearch.cleanUpData();
-
-            return Result.COMPLETED;
-        }
-    }
-
-    private final DockerOpenSearch elasticSearch;
-
-    public TestDockerOpenSearchModule(DockerOpenSearch elasticSearch) {
-        this.elasticSearch = elasticSearch;
+    @Override
+    public Statement apply(Statement base, Description description) {
+        return base;
     }
 
     @Override
-    protected void configure() {
-        bind(OpenSearchConfiguration.class).toInstance(elasticSearch.configuration());
-        Multibinder.newSetBinder(binder(), CleanupTasksPerformer.CleanupTask.class)
-            .addBinding()
-            .toInstance(new ESContainerCleanUp(elasticSearch));
+    public void await() {
+        openSearch.flushIndices();
     }
 
+    @Override
+    public Module getModule() {
+        return new TestDockerOpenSearchModule(openSearch);
+    }
+
+    public DockerOpenSearch getDockerEs() {
+        return openSearch;
+    }
+
+    public void start() {
+        openSearch.start();
+    }
 }
