+++ server/container/guice/opensearch/src/test/java/org/apache/james/TikaExtension.java
