+++ server/protocols/webadmin-integration-test/distributed-webadmin-integration-test/pom.xml
@@ -82,6 +82,13 @@
         </dependency>
         <dependency>
             <groupId>${james.groupId}</groupId>
+            <artifactId>james-server-guice-opensearch</artifactId>
+            <version>${project.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>${james.groupId}</groupId>
             <artifactId>james-server-webadmin-cassandra-data</artifactId>
             <scope>test</scope>
         </dependency>
